+++ README.md
@@ -1,30 +1,38 @@
 # Pretty-and-Pink-Space-Invaders
-A recreation of the classic *Space Invaders* Game using Pygame! This *Pretty-and-Pink* version features a gorgeous pink and purple colour scheme, epic sound effects (would reccommend full volume for maximum enjoyment) and dynamically moving cute aliens that get faster as the rounds progress. You need to defend well as the International Space Station is at stake. 
 
-This game was created by following *"The Art of Doing: Video Game Creation using Python and Pygame"* course by Michael Eramo. Excellent course! 
+A recreation of the classic *Space Invaders* Game using Pygame! This *Pretty-and-Pink* version features a gorgeous pink
+and purple colour scheme, epic sound effects (would recommend full volume for maximum enjoyment) and dynamically moving
+cute aliens that get faster as the rounds progress. You need to defend well as the International Space Station is at
+stake.
+
+This game was created by following *"The Art of Doing: Video Game Creation using Python and Pygame"* course by Michael
+Eramo. Excellent course!
 
 ## Playing ##
+
 1. Use the &#8592; key to move the Space Station left.
 2. Use the &#8594; key to move the Space Station right.
-3. Use the *space* key to shoot a bullet at the aliens.
+3. Use the `space` key to shoot a bullet at the aliens.
 
-If you get hit by an alien bullet or the aliens reach the station, you lose a life and the aliens will reset. You have 5 lives in total. If you shoot the oscillating love heart you will get a bonus life. If you clear all the aliens, you move onto the next round. 
+If you get hit by an alien bullet or the aliens reach the station, you lose a life and the aliens will reset. You have 5
+lives in total. If you shoot the oscillating love heart you will get a bonus life. If you clear all the aliens, you move
+onto the next round.
 
-The instantaneous accuracy is calculated everytime you shoot a bullet - you'll have to consider the accuracy and speed trade off when firing your bullets. 
+The instantaneous accuracy is calculated everytime you shoot a bullet - you'll have to consider the accuracy and speed
+trade off when firing your bullets.
 
-I wonder what the treasure chest contains? 
+I wonder what the treasure chest contains?
 
 ## Installing ## 
+
 1. Ensure *Python* and *Pygame* are installed on your computer.
-2. Download the *.wav*, *.png*, *.svg* files  into the same folder. I know the program is quite needy (just like me) and requires lots of files to run - but trust me, it'll be worth it.
-3. Update the working directory line (*line 14*) to the location of these files. 
-4. Run *space_invaders.py*
+2. Download all files into the same folder structure on your computer as this repo. I know the program is quite needy (
+   just like me) and requires lots of files to run - but trust me, it'll be worth it.
+3. Run `main.py`.
 
 ## Modifying ##
-Have fun and feel free to modify the parameters such as velocity or lives as you please. Screaming if you get hit by an alien is compulsory. 
 
-![Screenshot 2024-09-07 200347](https://github.com/user-attachments/assets/8d864d3b-e1c6-40f1-bd27-3057ad9912c7)
+Have fun and feel free to modify the parameters such as velocity or lives as you please. Screaming if you get hit by an
+alien is compulsory.
 
-
-
-
+![](screen_shot.png)
