+++ sounds/alien_hit.wav
